splitting out functionality into separate modules
