+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Changed
+* This project now uses `pandas` and parses a table-like file (CSV) _with_ header, to support upstream changes in data generation by `looptrace`.
+* Splitting old functionality out into separate modules
+
 ## [v0.1.0] - 2024-04-20
  
 ### Added
