+++ looptrace_loci_vis/__init__.py
@@ -1,3 +1,3 @@
 """Napari plugin for visualising locus-specific points from looptrace"""
 
-__version__ = "0.1.0"
+__version__ = "0.2dev"
